refactor: kafka consumers
