+++ docker-compose.yml
@@ -1,5 +1,3 @@
-version: "3"
-
 services:
   backend:
     build:
@@ -19,7 +17,20 @@ services:
       - ./.env:${DATA_FOLDER_PATH}/env-file
       - backend:${DATA_FOLDER_PATH}
 
-  faststream:
+  mapping-with-parsing-consumers:
+    image: datastep-backend
+    restart: always
+    command: bash -c "
+      cd /app/src
+      && faststream run --reload --host kafka:app"
+    volumes:
+      - ./src:/app/src
+      - ./.env:/app/.env
+      - backend:${DATA_FOLDER_PATH}
+    depends_on:
+      - backend
+
+  nsi-fetch-consumer:
     image: datastep-backend
     restart: always
     command: bash -c "
